TST Fix tests to work with latest numpy.

+++ mahotas/tests/test_watershed.py
@@ -48,8 +48,8 @@ def test_watershed2():
 
 def test_mix_types():
     f = np.zeros((64,64), np.uint16)
-    f += np.indices(f.shape)[1]**2
-    f += (np.indices(f.shape)[0]-23)**2
+    f += (np.indices(f.shape)[1]**2).astype(np.uint16)
+    f += ((np.indices(f.shape)[0]-23)**2).astype(np.uint16)
     markers = np.zeros((64,64), np.int64)
     markers[32,32] = 1
 # Below used to force a crash (at least in debug mode)
